+++ packages/fresh_graphql/lib/fresh_graphql.dart
@@ -4,6 +4,7 @@ export 'package:fresh/fresh.dart'
         InMemoryTokenStorage,
         OAuth2Token,
         RevokeTokenException,
+        Token,
         TokenHeaderBuilder,
         TokenStorage;
 export 'src/fresh_link.dart';